requirements.txt
django==2.1.2
djangorestframework==3.8.2
django-filter
pandas

django-heroku
gunicorn
